refactor : randomNumberGenerator을 util 폴더에 설정
