+++ __tests__/RandomNumberGenerator.test.js
@@ -1,4 +1,4 @@
-import randomNumberGenerator from "../src/domain/RandomNumberGenerator.js";
+import randomNumberGenerator from "../src/util/RandomNumberGenerator.js";
 import { systemSetting } from "../src/constants/systemSetting.js";
 test("randomNumberGenerator는 0~9 사이 수를 생성해야 한다", () => {
   //given
